+++ ldap4j-java/src/main/java/hu/gds/ldap4j/ldap/DerefAliases.java
@@ -14,7 +14,7 @@ public enum DerefAliases {
         this.value=value;
     }
 
-    public ByteBuffer write() throws Throwable {
+    public ByteBuffer write() {
         return DER.writeEnumeratedTag(value);
     }
 }